+++ CHANGES.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.2] - 2022-12-02
+### Added
+- It is now possible to modify references to external files (until now you could only delete and add them again).
+
+### Fixed
+- Importing external references has received a complete overhaul. The import now properly handles transitive dependencies and handles files which are included multiple times through different paths. It can now also handle functions modules and variables being moved between files.
+
+
 ## [0.5.1] - 2022-11-29
 ### Fixed
 - Spatially [arranging nodes](manual/manual.md#things-to-keep-in-mind-when-using-variables) to affect output order now also properly works for nested modules.